+++ mobile/android/android-components/docs/changelog.md
@@ -22,6 +22,14 @@ permalink: /changelog/
   * ⚠️ **This is a breaking change**: `alwaysAllowedSchemes` is removed as a parameter for `AppLinksInterceptor`.
   * Added `engineSupportedSchemes` as a parameter for `AppLinksInterceptor`.  This allows the caller to specify which protocol is supported by the engine.
     * Using this information, app links can decide if a protocol should be launched in a third party app or not regardless of user preference.
+    
+* **feature-sitepermissions**
+  * ⚠️ **This is a breaking change**: add parameters `autoplayAudible` and `autoplayInaudible` to `SitePermissionsRules`.
+    * This allows autoplay settings to be controlled for specific sites, rather than globally.
+  
+* **concept-engine**
+  * ⚠️ **This is a breaking change**: remove deprecated GeckoView setting `allowAutoplayMedia`
+    * This should now be controlled for individual sites via `SitePermissionsRules`
 
 # 33.0.0